% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_check_data.R
\name{extract_checks}
\alias{extract_checks}
\title{Parse and aggregate check results nested within a run block}
\usage{
extract_checks(doc)
}
\description{
Parse and aggregate check results nested within a run block
}
