% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_check_data.R
\name{extract_result}
\alias{extract_result}
\title{Map XML result properties into a single row data frame}
\usage{
extract_result(result)
}
\description{
Map XML result properties into a single row data frame
}
