% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_check_data.R
\name{extract_run}
\alias{extract_run}
\title{Extract root run metadata attributes from an XML document}
\usage{
extract_run(doc, filename)
}
\description{
Extract root run metadata attributes from an XML document
}
