% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_check_data.R
\name{extract_value}
\alias{extract_value}
\title{Extract a character text value from a nodeset using an XPath}
\usage{
extract_value(nodeset, xpath)
}
\description{
Extract a character text value from a nodeset using an XPath
}
