% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_flower.R
\name{fair_flower}
\alias{fair_flower}
\title{Extract Latest Metric State and Render a FAIR Flower Plot}
\usage{
fair_flower(df, title = NA, filename = NA)
}
\arguments{
\item{df}{data frame containing chronological monthly scores with columns \code{metric} and \code{mean}.}

\item{title}{title of plot}

\item{filename}{filename to save file to}
}
\value{
A polar ggplot object
}
\description{
Extract Latest Metric State and Render a FAIR Flower Plot
}
