% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_scores.R
\name{get_agg_data}
\alias{get_agg_data}
\title{Fetch Aggregated FAIR Assessment Scores}
\usage{
get_agg_data()
}
\value{
A data frame containing the aggregated metrics
}
\description{
Connects to the DataONE Quality API to fetch raw quality assessment scores
for suite: "FAIR-suite-0.5.0"
}
