% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_last_versions.R
\name{get_first_last_versions}
\alias{get_first_last_versions}
\title{Get first and last versions of a dataset}
\usage{
get_first_last_versions(data)
}
\arguments{
\item{data}{data frame from \code{get_agg_data}}
}
\value{
data.frame with \code{version_position} column added
}
\description{
Vectorized way to retrieve the first and last versions of a sequenceId. This
is considerably faster than doing it the dplyr way.
}
