% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_scores.R
\name{get_scores_cum_mean_ym}
\alias{get_scores_cum_mean_ym}
\title{Calculate Cumulative Monthly FAIR Scores}
\usage{
get_scores_cum_mean_ym(data)
}
\arguments{
\item{data}{A data frame containing at least the following columns:
\code{dateUploaded} , \code{scoreFindable}, \code{scoreAccessible},
\code{scoreInteroperable}, and \code{scoreReusable}.}
}
\value{
A data frame with columns: \code{ym}, \code{num_mean},
\code{num_cum_mean}, \code{metric} (Factor: Findable, Accessible,
Interoperable, Reusable), and \code{mean}.
}
\description{
Aggregates FAIR scores by month and calculates their cumulative mean over time,
returning a long-format data frame suitable for plotting. Filters out records
uploaded prior to 2000.
}
