% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_scores.R
\name{get_scores_monthly_state}
\alias{get_scores_monthly_state}
\title{Calculate Monthly State-Based FAIR Scores}
\usage{
get_scores_monthly_state(data)
}
\arguments{
\item{data}{A data frame containing at least the following columns:
\code{dateUploaded}, \code{sequenceId},  \code{scoreFindable}, \code{scoreAccessible},
\code{scoreInteroperable}, and \code{scoreReusable}.}
}
\value{
A data frame with columns: \code{ym} (Date), \code{num_mean} (integer count of active
sequences), \code{metric} (character: f, a, i, r), and \code{mean} (numeric score 0-100).
}
\description{
Evaluates the continuous state of repository FAIR scores over time. For each
unique sequence identifier, it selects the last upload per month, completes
a monthly time-grid, carries scores forward (LOCF) to account for persistent
datasets, and aggregates monthly means.
}
