% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_scores.R
\name{get_scores_quantiles_ym}
\alias{get_scores_quantiles_ym}
\title{Calculate Monthly FAIR Score Quantiles and Means}
\usage{
get_scores_quantiles_ym(data)
}
\arguments{
\item{data}{A data frame containing \code{dateUploaded}, \code{scoreFindable}, \code{scoreAccessible},
\code{scoreInteroperable}, and \code{scoreReusable}.}
}
\value{
A data frame with columns: \code{ym}, \code{metric},
\code{num_records}, \code{mean_score}, \code{p25},
and \code{p75} (numeric).
}
\description{
Calculate Monthly FAIR Score Quantiles and Means
}
