% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suite_summary.R
\name{get_suite_summary}
\alias{get_suite_summary}
\title{Get a summary of all the checks and their xpaths/jq}
\usage{
get_suite_summary(suite_path, checks_dir)
}
\arguments{
\item{suite_path}{}

\item{checks_dir}{}
}
\description{
Get a summary of all the checks and their xpaths/jq
}
