% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_flower.R
\name{make_flower_plot}
\alias{make_flower_plot}
\title{Generate a Polar FAIR Flower/Petal Diagram}
\usage{
make_flower_plot(.Data, title = NA)
}
\arguments{
\item{.Data}{A data frame containing individual metric rows with columns \code{score} and \code{label}.}

\item{title}{Character string title for the plot canvas. Defaults to \code{NA}.}
}
\value{
A theme-void polar coordinate ggplot visual asset with an absolute center score.
}
\description{
Derived from \code{fair_flower} from the \code{flowerplot} package, which I couldn't get to work
correctly and don't have time to maintain.
}
