% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_check_data.R
\name{process_xml_file}
\alias{process_xml_file}
\title{Parse single XML report into run and check blocks safely}
\usage{
process_xml_file(filename)
}
\description{
Parse single XML report into run and check blocks safely
}
