% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_check_data.R
\name{slice}
\alias{slice}
\title{Split up a vector into equal sized chunks}
\usage{
slice(input, by = 2)
}
\arguments{
\item{input}{An input vector.}

\item{by}{The length of the resulting vectors.}
}
\description{
Split up a vector into equal sized chunks
}
