% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aggregate_scores.R
\name{calc_aggregate_scores}
\alias{calc_aggregate_scores}
\title{Compute Aggregate FAIR Scores from Apache Arrow Repositories}
\usage{
calc_aggregate_scores(checks_dir, runs_dir)
}
\arguments{
\item{checks_dir}{Character string path to the partitioned checks parquet directory.}

\item{runs_dir}{Character string path to the partitioned execution runs parquet directory.}
}
\value{
A tidy, long-format data frame containing processed, wide-mapped FAIR dimensions.
}
\description{
This is probably not needed but might be nice if we don't trust the scorer
}
